requirements.txt
openai
python-pptx
python-dotenv
requests
rich
python-slugify
